% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverPercent.R
\name{coverPercent}
\alias{coverPercent}
\title{Computes coverage percentage of different land use classes from a raster image using buffers of different sizes.}
\usage{
coverPercent(Raster, surveySites, bufferSizes)
}
\arguments{
\item{Raster}{A categorical raster layer object , with projected CRS}

\item{surveySites}{Spatial points object with projected CRS. The first columm of its atribute table must contain the point identification}

\item{bufferSizes}{size in meters of the buffer to be created around each survey site}
}
\value{
data frame
}
\description{
Computes coverage percentage of different land use classes from a raster image using buffers of different sizes.
}
