% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverPercent.R
\name{coverPercent}
\alias{coverPercent}
\title{Computes coverage percentage of different land use classes from a raster image using moving windows.}
\usage{
coverPercent(landScape, landClass, windowSize, windowForm)
}
\arguments{
\item{landScape}{A categorical raster layer object}

\item{landClass}{Landuse class for which coverage percentage will be calculated}

\item{windowSize}{size, in units of the landScape CRS, of the moving window to be created}

\item{windowForm}{the type of moving window: "circle" , "Gauss", "rectangle"}
}
\value{
a SpatRaster image
}
\description{
Computes coverage percentage of different land use classes from a raster image using moving windows.
}
